Simple grouping with untyped properties or open types (#2034)
In the simple grouping algorithm, the last sub-bullet point of the
projection function says:
> Otherwise, if $p'$ is a prefix of a path in $G$, let $v[q]=s_G
(u[q],p')$

and the recursively called function $s_G$ assumes that its first
argument is a structured instance. But if $q$ is an untyped property or
a property of an open type, $u[q]$ can also be primitive<sup>1</sup>,
and then the first bullet point does not make sense:
> Let $v$ be an instance of the type of $u$ without properties and
without entity-id.

I suggest to rephrase the last sub-bullet point so that the recursive
call is omitted in such cases. This is the least invasive change and its
leads to $q$ being omitted from the projection.

<sup>1</sup> Of course $u[q]$ can also be primitive if $q$ declares a
primitive type, but the problem occurs only if $p'$ is a prefix of a
path in $G$, which means for a grouping criterion of the form $q/r$ or
$s/q/r$. And if $q$ declares a primitive type, $q/r$ does not make sense
at all.